requirements.txt
# Web frameworks
fastapi==0.103.1
flask==2.3.3
uvicorn==0.23.2
gunicorn==21.2.0

# Database
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7

# Authentication
pyjwt==2.8.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6

# Data processing
pandas==2.1.0
openpyxl==3.1.2
xlsxwriter==3.1.2

# Email
fastapi-mail==1.4.1

# Testing
pytest==7.4.2
httpx==0.24.1

# Utilities
python-dotenv==1.0.0
pydantic==2.3.0
email-validator==2.0.0
